requirements.txt
livekit-agents[deepgram,openai,cartesia,silero,turn-detector]
livekit-plugins-noise-cancellation
python-dotenv
sentence-transformers
numpy
openai
flask